+++ src/components/Posts/comment.js
@@ -29,8 +29,8 @@ export default function Comment(data) {
 
   const handleLike = () => {
     if (!data.liked){
-      comment_like(data.data.author.id, user, data.data.id, "context", like_success);
-      document.getElementsByClassName("heart")
+      console.log("Attempt post like for comment", data)
+      comment_like(data.data.author.id, user, postUrl, data.data.id, "context", like_success);
     }
     else {
       //delete like object
@@ -70,7 +70,7 @@ export default function Comment(data) {
               onClick={handleLike}
             >
               <svg version="1.1" id="heart-15" xmlns="http://www.w3.org/2000/svg" width="2.5em" height="2.5em" viewBox="0 0 15 15">
-                <path className="heart" fill={data.data.liked ? "red" : "var(--driftwood)"}
+                <path className="heart" fill={data.liked ? "red" : "white"}
                   d="M13.91,6.75c-1.17,2.25-4.3,5.31-6.07,6.94c-0.1903,0.1718-0.4797,0.1718-0.67,0C5.39,12.06,2.26,9,1.09,6.75&#xA;&#x9;C-1.48,1.8,5-1.5,7.5,3.45C10-1.5,16.48,1.8,13.91,6.75z"/>
               </svg>
             </button>
